% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_transect_interactive.R
\name{clip_transect_interactive}
\alias{clip_transect_interactive}
\title{Interactive Transect Clipping with lidR}
\usage{
clip_transect_interactive(las, width = 5, res = 1, subcircle = 0)
}
\arguments{
\item{las}{A character string specifying the path to a LAS/LAZ file or a LAS object.}

\item{width}{Numeric. The width of the transect. Default is 1.}

\item{res}{Numeric. The resolution for rasterizing the canopy height model. Default is 0.5.}

\item{subcircle}{Numeric. Parameter for the \code{p2r} function used in canopy rasterization.
Default is 0.}
}
\value{
A LAS object corresponding to the clipped transect.
}
\description{
This function interactively clips a transect from a LAS dataset. It first
generates a surface height model and displays it. The user is then prompted
to click two points (start and end of the transect) on the plot. A transect polygon
is constructed from these points and used to clip the LAS dataset.
}
\examples{
\dontrun{
  # use the Megaplot example dataset from lidR:
  lasFile <- system.file("extdata", "Megaplot.laz", package = "lidR")

  las_clipped <- clip_transect_interactive(
      las       = lasFile,
      width     = 5,
      res       = 1,
      subcircle = 0.25
    )
    }
}
